requirements.txt
asgiref==3.3.1
cryptography==3.3.1
Django==3.1.4
django-allauth==0.44.0
django-appconf==1.0.4
django-countries==7.0
django-crispy-forms==1.10.0
django-localflavor==3.0.1
django-user-agents==0.4.0
oauthlib==3.1.0
Pillow==8.1.0
PyJWT==2.0.0
python-stdnum==1.16
python3-openid==3.2.0
pytz==2020.5
requests-oauthlib==1.3.0
shortuuid==1.0.1
sqlparse==0.4.1
stripe==2.55.1
ua-parser==0.10.0
user-agents==2.2.0
